Move remaining comment action components into dropdowns/comments
